ci: add gradle/gradle-build-action step
This step adds caching of Gradle dependencies to the pipelines. This
should make the builds a bit faster. Only builds on the main branch
write to the cache. But cache entries will be read by all other
branches [1].

This action also stops all running Gradle daemons [2]. Therefore it is
fine to run the Gradle CLI without the --no-daemon flag. This will also
improve the performance if the pipeline contains multiple Gradle steps.

[1] https://github.com/gradle/gradle-build-action#select-branches-that-should-write-to-the-cache
[2] https://github.com/gradle/gradle-build-action#stopping-the-gradle-daemon